+++ fairlearn/preprocessing/_correlation_remover.py
@@ -2,12 +2,13 @@
 # Licensed under the MIT License.
 
 from collections.abc import Iterable
-from test.unit.fixes import validate_data
 
 import numpy as np
 import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
+
+from fairlearn.utils._fixes import validate_data
 
 
 class CorrelationRemover(TransformerMixin, BaseEstimator):
